requirements.txt
# Non RobotPy packages that we use
black
pylint==2.15.3
pytest
wheel
# RobotPy packages.  We explicitly list the version number on all these packages so that when RobotPy
# updates we don't break the world until we have a chance to test/update.
pyfrc==2023.0.0b4
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2023.0.0b7
robotpy-commands-v2==2023.0.0b7
robotpy-ctre==2023.0.0b1
robotpy-hal==2023.0.0b7
robotpy-halsim-gui==2023.0.0b7
robotpy-installer==2023.0.0b5
robotpy-navx==2023.0.0b1
robotpy-wpimath==2023.0.0b7
robotpy-wpiutil==2023.0.0b7
wpilib==2023.0.0b7

